# api/openapi.yaml
openapi: 3.0.3
info:
  title: JobTracker API
  description: Self-hosted job application tracker with AI assistance
  version: 0.2.0

servers:
  - url: http://localhost:8080/api/v1

paths:
  /applications:
    get:
      operationId: listApplications
      summary: List applications, optionally filtered by status
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      summary: Create a new application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"

  /applications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: getApplication
      summary: Get a single application
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateApplication
      summary: Partially update an application (e.g. status change)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteApplication
      summary: Delete an application
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  /profile:
    get:
      operationId: getProfile
      summary: Get the stored CV profile
      responses:
        "200":
          description: Stored profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateProfile
      summary: Create or replace the CV profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Profile"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/BadRequest"

  /ai/parse-job:
    post:
      operationId: parseJobPosting
      summary: Parse a raw job posting text into a structured application draft
      description: >
        Sends the raw posting text to the LLM and returns a draft Application.
        Nothing is persisted; the client reviews the draft and calls
        POST /applications to save it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [raw_text]
              properties:
                raw_text:
                  type: string
                  description: Full text of the job posting (copy-pasted)
                url:
                  type: string
                  description: Optional source URL
      responses:
        "200":
          description: Parsed draft (not persisted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationInput"
        "422":
          description: Text could not be parsed as a job posting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /ai/parse-url:
    post:
      operationId: parseJobURL
      summary: Fetch a public job posting URL and parse it
      description: >
        Works for public pages (company career sites). Pages behind login
        walls (e.g. LinkedIn) return 422 with a hint to copy-paste instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
      responses:
        "200":
          description: Parsed draft (not persisted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationInput"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: Page could not be read or parsed as a job posting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: URL could not be fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /ai/score:
    post:
      operationId: scoreApplication
      summary: Score the stored CV against one application's job description
      description: >
        Loads the stored CV (PUT /profile) and the application's job_description
        from the database, asks the LLM for a fit analysis, persists the result
        on the application (fit_score, score_details) and returns it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreRequest"
      responses:
        "200":
          description: Scoring result (also persisted on the application)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /ai/cover-letter:
    post:
      operationId: generateCoverLetter
      summary: Generate a tailored cover letter for one application
      description: >
        Uses the stored CV and the application's job_description. The result is
        returned only (not persisted) — the user reviews and edits it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoverLetterRequest"
      responses:
        "200":
          description: Generated cover letter draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /inbox/sync:
    post:
      operationId: syncInbox
      summary: Fetch and classify recent emails now
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                required: [fetched, new_events]
                properties:
                  fetched:
                    type: integer
                    description: Messages examined in this run
                  new_events:
                    type: integer
                    description: New inbox events created
        "503":
          description: Gmail not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /inbox/events:
    get:
      operationId: listInboxEvents
      summary: List classified email events, newest first
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/InboxEventKind"
        - name: include_dismissed
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboxEvent"

  /inbox/events/{id}/apply:
    post:
      operationId: applyInboxEvent
      summary: Apply the suggested status to the linked application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Event has no linked application or no suggested status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /inbox/events/{id}/dismiss:
    post:
      operationId: dismissInboxEvent
      summary: Hide an event from the default list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Dismissed
        "404":
          $ref: "#/components/responses/NotFound"
components:
  schemas:
    ApplicationStatus:
      type: string
      enum: [saved, applied, interview, offer, rejected, ghosted]
    InboxEventKind:
      type: string
      enum: [job_alert, application_update, irrelevant]

    InboxEvent:
      type: object
      required:
        [
          id,
          gmail_message_id,
          received_at,
          from,
          subject,
          kind,
          summary,
          dismissed,
        ]
      properties:
        id:
          type: integer
          format: int64
        gmail_message_id:
          type: string
        received_at:
          type: string
          format: date-time
        from:
          type: string
        subject:
          type: string
        kind:
          $ref: "#/components/schemas/InboxEventKind"
        summary:
          type: string
          description: AI-written 1-3 sentence summary (job matches or update meaning)
        application_id:
          type: integer
          format: int64
          nullable: true
        suggested_status:
          $ref: "#/components/schemas/ApplicationStatus"
        confidence:
          type: number
          format: double
          nullable: true
        dismissed:
          type: boolean

    SalaryPeriod:
      type: string
      enum: [hourly, monthly, yearly]

    EmploymentType:
      type: string
      enum: [werkstudent, fulltime, parttime, internship]

    ApplicationInput:
      type: object
      required: [company, position]
      properties:
        company:
          type: string
        position:
          type: string
        city:
          type: string
        source:
          type: string
          enum: [linkedin, indeed, stepstone, referral, company_site, other]
        url:
          type: string
        employment_type:
          $ref: "#/components/schemas/EmploymentType"
        salary_min:
          type: number
          format: double
          nullable: true
        salary_max:
          type: number
          format: double
          nullable: true
        salary_period:
          $ref: "#/components/schemas/SalaryPeriod"
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        applied_at:
          type: string
          format: date
          nullable: true
        notes:
          type: string
        job_description:
          type: string
          description: Raw posting text, reused for CV scoring and cover letters

    ApplicationUpdate:
      type: object
      description: All fields optional; only provided fields are updated
      properties:
        company:
          type: string
        position:
          type: string
        city:
          type: string
        source:
          type: string
          enum: [linkedin, indeed, stepstone, referral, company_site, other]
        url:
          type: string
        employment_type:
          $ref: "#/components/schemas/EmploymentType"
        salary_min:
          type: number
          format: double
          nullable: true
        salary_max:
          type: number
          format: double
          nullable: true
        salary_period:
          $ref: "#/components/schemas/SalaryPeriod"
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        applied_at:
          type: string
          format: date
          nullable: true
        notes:
          type: string
        job_description:
          type: string

    Application:
      allOf:
        - $ref: "#/components/schemas/ApplicationInput"
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id:
              type: integer
              format: int64
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            fit_score:
              type: integer
              nullable: true
              description: AI-generated match score (0-100), set by POST /ai/score
            score_details:
              type: string
              nullable: true
              description: JSON string with matched/missing keywords and suggestions

    Profile:
      type: object
      required: [cv_text]
      properties:
        cv_text:
          type: string
          description: Raw text of the user's CV
        updated_at:
          type: string
          format: date-time
          readOnly: true

    ScoreRequest:
      type: object
      required: [application_id]
      properties:
        application_id:
          type: integer
          format: int64

    ScoreResponse:
      type: object
      required: [score, matched_keywords, missing_keywords, suggestions]
      properties:
        score:
          type: integer
          description: Overall fit score, 0-100
        matched_keywords:
          type: array
          items:
            type: string
        missing_keywords:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string

    CoverLetterRequest:
      type: object
      required: [application_id, language, tone]
      properties:
        application_id:
          type: integer
          format: int64
        language:
          type: string
          enum: [en, de, tr]
        tone:
          type: string
          enum: [formal, warm, concise]

    CoverLetterResponse:
      type: object
      required: [cover_letter]
      properties:
        cover_letter:
          type: string

    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string

  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"